+++ skore-ui/src/components/MediaWidgetSelector.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import CrossValidationReport from "@/components/CrossValidationReport.vue";
 import DataFrameWidget from "@/components/DataFrameWidget.vue";
 import HtmlSnippetWidget from "@/components/HtmlSnippetWidget.vue";
 import ImageWidget from "@/components/ImageWidget.vue";
@@ -39,4 +40,11 @@ function matchMediaType(mediaType: string) {
     :src="props.item.data"
   />
   <HtmlSnippetWidget v-if="matchMediaType('text/html')" :src="props.item.data" />
+  <CrossValidationReport
+    v-if="matchMediaType('application/vnd.skore.cross_validation')"
+    :scalar-results="props.item.data.scalar_results"
+    :tabular-results="props.item.data.tabular_results"
+    :plots="props.item.data.plots"
+    :sections="props.item.data.sections"
+  />
 </template>
